move week select to own file
